+++ src/main/java/com/estsoft13/matdori/service/UserDetailService.java
@@ -6,7 +6,6 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-
 @Service
 public class UserDetailService implements UserDetailsService {
 
@@ -23,3 +22,4 @@ public class UserDetailService implements UserDetailsService {
                 .orElseThrow(() -> new UsernameNotFoundException(email));
     }
 }
+
