+++ src/main/resources/templates/meetingSearch.html
@@ -7,6 +7,7 @@
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto&display=swap">
     <link rel="stylesheet" href="/css/normalize.css">
     <link rel="stylesheet" href="/css/mainMeeting.css">
+    <title>미팅 검색 결과</title>
 </head>
 <body>
 <header>
